pyproject.toml
# pyproject.toml
[tool.poetry]
name = "Trip-Planner"
version = "1.0.0"
description = "The best trip planner on Planet Earth."
authors = ["Bharat Sharma <bharat8717sharma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
langchain = "^0.1.20"
langgraph = "^0.0.48"
langchain-google-genai = "^1.0.3"
langchain-openai = "^0.1.6"
langchain-groq = "^0.1.3"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
streamlit = "^1.35.0"
pyyaml = "^6.0.1"
sqlmodel = "^0.0.18"
langchainhub = "^0.1.15"                                 # For pulling prompts

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
